[refactor]: string constants
