+++ lib/blocs/auth/register/register_cubit.dart
@@ -1,4 +1,5 @@
 import 'package:bloc/bloc.dart';
+import 'package:dalal_street_client/constants/error_messages.dart';
 import 'package:dalal_street_client/grpc/client.dart';
 import 'package:dalal_street_client/main.dart';
 import 'package:dalal_street_client/proto_build/actions/Register.pb.dart';
@@ -31,7 +32,7 @@ class RegisterCubit extends Cubit<RegisterState> {
       }
     } catch (e) {
       logger.e(e);
-      emit(const RegisterFailure('Failed to reach server. Try again later'));
+      emit(const RegisterFailure(failedToReachServer));
     }
   }
 }
