+++ lib/blocs/auth/verify_phone/enter_otp/enter_otp_cubit.dart
@@ -1,5 +1,6 @@
 import 'package:bloc/bloc.dart';
 import 'package:dalal_street_client/blocs/user/user_bloc.dart';
+import 'package:dalal_street_client/constants/error_messages.dart';
 import 'package:dalal_street_client/grpc/client.dart';
 import 'package:dalal_street_client/main.dart';
 import 'package:dalal_street_client/proto_build/actions/AddPhone.pb.dart';
@@ -30,7 +31,7 @@ class EnterOtpCubit extends Cubit<OtpState> {
       }
     } catch (e) {
       logger.e(e);
-      emit(const OtpFailure('Failed to reach server. Try again later'));
+      emit(const OtpFailure(failedToReachServer));
     }
     emit(OtpInitial(phone));
   }
@@ -54,7 +55,7 @@ class EnterOtpCubit extends Cubit<OtpState> {
       }
     } catch (e) {
       logger.e(e);
-      emit(const OtpFailure('Failed to reach server. Try again later'));
+      emit(const OtpFailure(failedToReachServer));
       emit(OtpInitial(phone));
     }
   }
