+++ lib/constants/app_info.dart
