+++ lib/constants/error_messages.dart
@@ -1,0 +1,1 @@
+const failedToReachServer = 'Failed to reach server. Try again later';
